% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-local-power.R
\name{check_equal_length}
\alias{check_equal_length}
\title{Check whether list elements have equal length}
\usage{
check_equal_length(lst)
}
\arguments{
\item{lst}{a list.}
}
\value{
logical scalar.
}
\description{
Check whether list elements have equal length
}
