% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-local-power.R
\name{check_identical}
\alias{check_identical}
\title{Check whether list elements are identical}
\usage{
check_identical(lst)
}
\arguments{
\item{lst}{a list.}
}
\value{
logical scalar.
}
\description{
Check whether list elements are identical
}
