% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{coef.model_fit}
\alias{coef.model_fit}
\title{Estimated coefficients from a model fit}
\usage{
\method{coef}{model_fit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{model_fit}.}

\item{...}{unused; retained for S3 method consistency.}
}
\value{
numeric vector of estimated parameters.
}
\description{
Estimated coefficients from a model fit
}
