% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage-gls.R
\name{fit_gls}
\alias{fit_gls}
\title{Fit a GLS model to first-stage estimates}
\usage{
fit_gls(m_tilde, Sigma, mean_fn, jac_mean_fn, start, method = "L-BFGS-B")
}
\arguments{
\item{m_tilde}{(numeric) Stacked mean vector from the first-stage estimates.}

\item{Sigma}{(matrix) Covariance matrix from the first-stage estimates.}

\item{mean_fn}{(function) that takes a vector of parameters (\code{gamma}) and
returns the stacked mean vector. This argument defines the model.}

\item{jac_mean_fn}{(function) that takes a vector of parameters (\code{gamma}) and
returns the Jacobian of the mean function. This is used to compute the
gradient of the GLS criterion function.}

\item{start}{(numeric) Starting values for the parameters to be estimated.}

\item{method}{(character) Optimization method to be used in \code{optim()}.
Default is \code{"BFGS"}.}
}
\value{
(list) A list containing:
\itemize{
\item \code{gamma_hat}: Estimated parameters (vector).
\item \code{criterion}: Value of the GLS criterion function at the optimum.
\item \code{optim}: Raw output from the \code{optim()} function.
}
}
\description{
\code{fit_gls()} fits a generalized least squares (GLS) model to the first-stage
estimates \code{m_tilde} and \code{Sigma}. The model is specified by the mean function
\code{mean_fn} and its Jacobian \code{jac_mean_fn}. The optimization is performed using
the \code{optim()} function, starting from the initial parameter values provided
in \code{start}. The optimization method can be specified using the \code{method}
argument.
}
