% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage-gls.R
\name{fit_gls_est_names}
\alias{fit_gls_est_names}
\title{Process GLS fit estimates and assign names based on the model structure}
\usage{
fit_gls_est_names(gamma_hat, null_model, slowing_models)
}
\arguments{
\item{gamma_hat}{(numeric) Vector of estimated parameters from the GLS fit.}

\item{null_model}{(boolean) Indicates whether the model is a null model
(TRUE) where the treatment effect is zero or a full model (FALSE).}

\item{slowing_models}{(list) A list of slowing model specifications for each
outcome, as returned by \code{build_slowing_models()}.}
}
\value{
(named numeric) A named vector of estimated parameters, where the
names are constructed based on the model structure and outcome indices.
}
\description{
Process GLS fit estimates and assign names based on the model structure
}
