% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage-gls.R
\name{fit_gls_split_gamma_hat}
\alias{fit_gls_split_gamma_hat}
\title{Split Estimated Slowing Parameters}
\usage{
fit_gls_split_gamma_hat(gamma_hat, null_model, working_model)
}
\arguments{
\item{gamma_hat}{(numeric) Vector of estimated parameters from the GLS fit.}

\item{null_model}{(boolean) Indicates whether the model is a null model
(TRUE) where the treatment effect is zero or a full model (FALSE).}
}
\value{
(list) A list containing two elements:
\itemize{
\item \code{gamma0_hat}: A list of estimated reference-trajectory parameters for each outcome.
\item \code{gamma1_hat}: A list of estimated slowing parameters for each outcome.
}
}
\description{
This function takes a vector of estimated parameters (\code{gamma_hat}) from a GLS
fit and splits it into two lists: one for the estimated reference-trajectory
parameters (\code{gamma0_hat}) and another for the estimated slowing parameters
(\code{gamma1_hat}). The splitting is based on the structure of the provided
\code{slowing_models} and whether the model is a null model or not.
}
