% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage-gls.R
\name{gls_criterion}
\alias{gls_criterion}
\title{\code{gls_criterion()} evaluates the GLS criterion function
Q(gamma) = (m_tilde - mu(gamma))' Sigma^{-1} (m_tilde - mu(gamma)).}
\usage{
gls_criterion(gamma, mean_fn, m_tilde, Sigma_inv)
}
\arguments{
\item{gamma}{(numeric) Vector of parameters for the mean function.}

\item{mean_fn}{(function) that takes a vector of parameters (\code{gamma}) and
returns the stacked mean vector.}

\item{m_tilde}{(numeric) Stacked mean vector from the first-stage estimates.}

\item{Sigma_inv}{(matrix) Inverse of the first-stage covariance matrix.}
}
\value{
(numeric) The value of the GLS criterion function evaluated at \code{gamma}.
}
\description{
The model is specified by the mean function \code{mean_fn}, which takes a vector
of parameters \code{gamma} and returns the stacked mean vector. The first-stage
estimates are given by \code{m_tilde} and \code{Sigma_inv} is the inverse of the
first-stage covariance matrix.
}
