% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4PL.R
\name{jacobian_4PL_asympt}
\alias{jacobian_4PL_asympt}
\title{Jacobian of the 4PL reference trajectory with free asymptotes}
\usage{
jacobian_4PL_asympt(times, params)
}
\arguments{
\item{times}{numeric vector of time points.}

\item{params}{numeric vector of length 4:
\code{c(inflection, slope, upper_asymptote, lower_asymptote)}.}
}
\value{
numeric matrix with \code{length(times)} rows and 4 columns (partial
derivatives with respect to \code{inflection}, \code{slope}, \code{upper_asymptote}, and
\code{lower_asymptote}).
}
\description{
Jacobian of the 4PL reference trajectory with free asymptotes
}
