% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_for_model_params}
\alias{model_for_model_params}
\title{Compose a model with its reduced-parameter model}
\usage{
model_for_model_params(working_model, model_for_params)
}
\arguments{
\item{working_model}{an object of class \code{model}.}

\item{model_for_params}{an object of class \code{model} mapping the reduced
(e.g. shared) parameter vector onto \code{working_model}'s full parameter
vector.}
}
\value{
an object of class \code{model} parameterized by
\code{model_for_params}'s parameters.
}
\description{
Compose a model with its reduced-parameter model
}
\keyword{internal}
