% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{new_model}
\alias{new_model}
\title{Finalize a working-model object}
\usage{
new_model(x = list())
}
\arguments{
\item{x}{list with elements \code{mean_fn}, \code{jacobian_fn},
\code{nuisance_params_position}, and \code{treatment_params_null}.}
}
\value{
an object of class \code{model}.
}
\description{
Given a list with a mean function, Jacobian function, and information about
which parameters are nuisance vs. treatment-effect parameters, computes the
derived null/treatment mean and Jacobian functions and assigns the \code{model}
class.
}
