% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage-gls.R
\name{plot_gls_fitted}
\alias{plot_gls_fitted}
\title{Plot GLS-fitted values against first-stage estimates}
\usage{
plot_gls_fitted(
  gls_fitted,
  treatment_strata = NULL,
  outcome_strata = NULL,
  times = NULL
)
}
\arguments{
\item{gls_fitted}{an object of class \code{model_fit} (see \code{\link[=two_stage_gls_null]{two_stage_gls_null()}}
or \code{\link[=two_stage_gls_full]{two_stage_gls_full()}}).}

\item{treatment_strata}{optional character/factor vector labeling each
element of \code{m_tilde} by treatment arm.}

\item{outcome_strata}{optional character/factor vector labeling each
element of \code{m_tilde} by outcome.}

\item{times}{optional numeric vector of time points for each element of
\code{m_tilde}.}
}
\value{
a \code{ggplot} object.
}
\description{
Requires the \code{ggplot2} package.
}
