% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print.model}
\alias{print.model}
\title{Print a working-model object}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{an object of class \code{model}.}

\item{...}{unused; retained for S3 method consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a working-model object
}
