% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{print.model_fit}
\alias{print.model_fit}
\title{Print a model fit}
\usage{
\method{print}{model_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{model_fit}.}

\item{...}{unused; retained for S3 method consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a model fit
}
