% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{print.summary.model_fit}
\alias{print.summary.model_fit}
\title{Print a model fit summary}
\usage{
\method{print}{summary.model_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.model_fit}.}

\item{...}{unused; retained for S3 method consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a model fit summary
}
