% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-local-power.R
\name{ref_d}
\alias{ref_d}
\title{Time-derivative of a reference trajectory, dispatched by model type}
\usage{
ref_d(times, ref = "4PL", params, ...)
}
\arguments{
\item{times}{numeric vector of measurement times.}

\item{ref}{character, the reference trajectory (\code{"4PL"} or \code{"nc_spline"}).}

\item{params}{numeric vector of reference-trajectory parameters.}

\item{...}{additional arguments forwarded to the spline helpers.}
}
\value{
numeric vector, the time-derivative of the reference trajectory.
}
\description{
Time-derivative of a reference trajectory, dispatched by model type
}
