% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{split_model}
\alias{split_model}
\title{Split a working model by parameter/mean-vector stratum}
\usage{
split_model(working_model, split_indices_mu, split_indices_params)
}
\arguments{
\item{working_model}{an object of class \code{model}.}

\item{split_indices_mu}{integer vector, the stratum of each element of the
mean vector.}

\item{split_indices_params}{list of integers, the stratum of each
parameter (elements may have length greater than one for shared-parameter
models).}
}
\value{
list of objects of class \code{model}, one per unique stratum.
}
\description{
Splits \code{working_model} into one sub-model per unique value of
\code{split_indices_params}, restricting the mean and Jacobian functions to the
corresponding rows/columns.
}
