% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted-test.R
\name{targeted_test}
\alias{targeted_test}
\title{Data-Adaptive Targeted Test Based on Slowing Model}
\usage{
targeted_test(
  m_tilde,
  Sigma,
  working_model,
  A,
  start,
  ols = FALSE,
  split_indices_mu = NULL,
  split_indices_params = NULL
)
}
\arguments{
\item{m_tilde}{(numeric) Stacked mean vector from the first-stage estimates.}

\item{Sigma}{(matrix) Covariance matrix from the first-stage estimates.}

\item{start}{(numeric) Starting values for the parameters to be estimated.}

\item{split_indices_mu}{integer vector where each element denotes the
stratum of the the corresponding element in \code{m_tilde}.}

\item{split_indices_params}{list of integers where each element denotes the stratum of the
the corresponding parameter in parameter vector for \code{working_model}. The
list elements can have length greater than one, which may be required for
shared-parameter models.}

\item{shared}{(boolean) If TRUE, the slowing model parameters are shared
across outcomes; if FALSE, they are outcome-specific.}
}
\value{
(list) A list containing the following elements:
\itemize{
\item \code{B_n}: The plug-in contrast matrix.
\item \code{gls_fitted_null}: The fitted model under the null hypothesis.
\item \code{statistic}: The computed test statistic.
\item \code{df}: The degrees of freedom for the test statistic.
\item \code{p_value}: The p-value associated with the test statistic.
}
}
\description{
\code{targeted_test()} estimates the slowing model parameters
under the null hypothesis and computes a targeted test statistic for the
omnibus null hypothesis of no treatment effect across multiple outcomes.
}
