% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted-test.R
\name{targeted_test_statistic}
\alias{targeted_test_statistic}
\title{Compute Chi-Squared Test Statistic}
\usage{
targeted_test_statistic(B, m_tilde, Sigma)
}
\arguments{
\item{B}{(matrix) The contrast matrix.}

\item{m_tilde}{(numeric) Stacked mean vector from the first-stage estimates.}

\item{Sigma}{(matrix) Covariance matrix from the first-stage estimates.}
}
\value{
(list) A list containing the following elements:
\itemize{
\item \code{statistic}: The computed test statistic.
\item \code{df}: The degrees of freedom for the test statistic.
\item \code{p_value}: The p-value associated with the test statistic.
}
}
\description{
Compute Chi-Squared Test Statistic
}
