% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage-gls.R
\name{two_stage_gls_null}
\alias{two_stage_gls_null}
\alias{two_stage_gls_full}
\title{Fit a slowing model using GLS}
\usage{
two_stage_gls_null(
  m_tilde,
  Sigma,
  working_model,
  start,
  ols = FALSE,
  split_indices_mu = NULL,
  split_indices_params = NULL
)

two_stage_gls_full(
  m_tilde,
  Sigma,
  working_model,
  start,
  ols = FALSE,
  split_indices_mu = NULL,
  split_indices_params = NULL
)
}
\arguments{
\item{m_tilde}{(numeric) Stacked mean vector from the first-stage estimates.}

\item{Sigma}{(matrix) Covariance matrix from the first-stage estimates.}

\item{start}{(numeric) Starting values for the parameters to be estimated.}

\item{split_indices_mu}{integer vector where each element denotes the
stratum of the the corresponding element in \code{m_tilde}.}

\item{split_indices_params}{list of integers where each element denotes the stratum of the
the corresponding parameter in parameter vector for \code{working_model}. The
list elements can have length greater than one, which may be required for
shared-parameter models.}
}
\value{
an object of class \code{model_fit} (see \code{\link[=new_model_fit]{new_model_fit()}}).
}
\description{
\code{two_stage_gls_full()} fits a generalized least squares (GLS) model to the
first-stage estimates. \code{two_stage_gls_null()} fits the same model under the
null of no treatment effect. It constructs the mean function based on the
provided slowing models, and then optimizes the GLS criterion function to
estimate the reference-trajectory parameters (and slowing parameters for
\code{two_stage_gls_full()}).
}
