% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{vcov.model_fit}
\alias{vcov.model_fit}
\title{Estimated covariance matrix from a model fit}
\usage{
\method{vcov}{model_fit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{model_fit}.}

\item{...}{unused; retained for S3 method consistency.}
}
\value{
numeric matrix, the estimated covariance matrix of the
coefficients.
}
\description{
Estimated covariance matrix from a model fit
}
